+++ src/main/java/roomescape/controller/ReservationController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import roomescape.ReservationDto;
+import roomescape.dto.ReservationDto;
 import roomescape.entity.Reservation;
 import roomescape.repository.ReservationRepository;
