+++ src/main/java/roomescape/dto/ReservationDto.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDate;
