+++ src/main/java/roomescape/dto/ReservationTimeDto.java
@@ -1,21 +1,22 @@
-package roomescape;
+package roomescape.dto;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
 import roomescape.entity.ReservationTime;
 
-public class TimeDto { // TODO: change class name
+public class ReservationTimeDto {
 
     private final Long id;
     private final LocalTime startAt;
 
-    public TimeDto(Long id, LocalTime startAt) {
+    public ReservationTimeDto(Long id, LocalTime startAt) {
         this.id = id;
         this.startAt = startAt;
     }
 
-    public static TimeDto from(ReservationTime time) {
+    public static ReservationTimeDto from(ReservationTime time) {
         // TODO: entity에서 dto로 변환하는 게 나을까?
-        return new TimeDto(time.getId(), time.getStartAt());
+        return new ReservationTimeDto(time.getId(), time.getStartAt());
     }
 
     public ReservationTime toEntity() {
